+++ packages/common/src/rule-engine/operators/types.ts
@@ -7,6 +7,7 @@ import {
   PrimitiveSchema,
   IdvCheckSchema,
   UboMismatchSchema,
+  AmlCheckV2Schema,
 } from '@/rule-engine/operators/schemas';
 
 import { OPERATION, OPERATOR } from './enums';
@@ -22,6 +23,8 @@ export type BetweenParams = z.infer<typeof BetweenSchema>;
 export type LastYearsParams = z.infer<typeof LastYearsSchema>;
 
 export type AmlCheckParams = z.infer<typeof AmlCheckSchema>;
+
+export type AmlCheckV2Params = z.infer<typeof AmlCheckV2Schema>;
 
 export type IdvCheckParams = z.infer<typeof IdvCheckSchema>;