+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/DocumentField/DocumentField.tsx
@@ -1,4 +1,4 @@
-import { AnyObject, ctw } from '@/common';
+import { ALLOWED_DOCUMENT_FILE_EXTENSIONS, AnyObject, ctw } from '@/common';
 import { IHttpParams, useHttp } from '@/common/hooks/useHttp';
 import { Button } from '@/components/atoms';
 import { Input } from '@/components/atoms/Input';
@@ -74,7 +74,8 @@ export const DocumentField: TDynamicFormField<IDocumentFieldParams> = ({ element
   );
 
   const { params } = element;
-  const { placeholder = 'Choose file', acceptFileFormats = undefined } = params || {};
+  const { placeholder = 'Choose file', acceptFileFormats = ALLOWED_DOCUMENT_FILE_EXTENSIONS } =
+    params || {};
   const { removeTask, getTaskById, isRunning } = useTaskRunner();
   const { documentState, updateState } = useDocumentState(
     element as IFormElement<'documentfield', IDocumentFieldParams>,