+++ services/workflows-service/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - fed1494bf4a7dde81915d4593765e7d89104c164: Bump
 - Updated dependencies [fed1494bf4a7dde81915d4593765e7d89104c164]
+- version bump
+- Updated dependencies
   - @ballerine/common@0.9.96
   - @ballerine/workflow-core@0.6.119
   - @ballerine/workflow-node-sdk@0.6.119