+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ballerine/workflows-service
 
+## 0.7.128
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.98
+  - @ballerine/workflow-core@0.6.121
+  - @ballerine/workflow-node-sdk@0.6.121
+
 ## 0.7.127
 
 ### Patch Changes